% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EXRQ.R
\name{qpareto}
\alias{qpareto}
\title{Quantile of the Pareto Distribution}
\usage{
qpareto(p, gamma)
}
\arguments{
\item{p}{the quantile level}

\item{gamma}{the shape parameter}
}
\value{
the pth quantile
}
\description{
Quantile of the Pareto Distribution
}

