% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EXRQ.R
\name{rpareto}
\alias{rpareto}
\title{Random Generation for the Pareto Distribution}
\usage{
rpareto(n, gamma)
}
\arguments{
\item{n}{number of observations}

\item{gamma}{the shape parameter}
}
\value{
a vector of n i.i.d. random variables from the Pareto distribution
}
\description{
Random Generation for the Pareto Distribution
}

